+++ pkg/api/cloud/stack/start.go
@@ -5,7 +5,7 @@ import (
 )
 
 // Start starts a cloud stack on a given server.
-func (s *Client) Start(ctx context.Context, request StopStartRestartRequest) (response StartStopRestartResponse, err error) {
+func (s *Client) Start(ctx context.Context, request StopStartRestartRequest) (response ActionResponse, err error) {
 	uri := "cloud/stack/start.json"
 	return s.stopStartRestart(ctx, request, uri)
 }
